Struct bitgrin_wallet_impls::HTTPNodeClient

pub struct HTTPNodeClient { /* fields omitted */ }

Methods

impl HTTPNodeClient

pub fn new(node_url: &str, node_api_secret: Option<String>) -> HTTPNodeClient

Create a new client that will communicate with the given grin node

pub fn chain_height(&self) -> Result<u64, Error>

Allow returning the chain height without needing a wallet instantiated