Module bitgrin_wallet_impls::test_framework

Structs

LocalWalletClient

WalletProxy
communicates with a chain instance or other wallet listener APIs via message queues

Functions

add_block_with_reward
Adds a block with a given reward to the chain and mines it

award_block_to_wallet
adds a reward output to a wallet, includes that reward in a block, mines the block and adds it to the chain, with option transactions included. Helpful for building up precise wallet balances for testing.

award_blocks_to_wallet
Award a blocks to a wallet directly

create_wallet
dispatch a db wallet

send_to_dest
send an amount to a destination

wallet_info
get wallet info totals